Refactor tests so that the stx tests also run in NodeJS
It would be nice to have tests split up into modules, but this causes
all kinds of headaches with running the tests in both the browser and
NodeJS. So we're keeping them in one file for now.
